requirements.txt
PyMuPDF==1.23.8
python-docx==1.1.0
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.5.1
langchain==0.1.12
rich==13.7.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
httpx==0.27.0 